Moved 13 tests from GeneralWorkflowIT to ManualPublishWorkflowIT
